+++ phpBB2/admin/xs_include_import2.php
@@ -90,10 +90,6 @@ if(!$f)
 	}
 	xs_error($lang['xs_error_cannot_open'] . '<br /><br />' . $lang['xs_import_back']);
 }
-if($write_local)
-{
-	$write_local_dir .= $header['template'] . '/';
-}
 fseek($f, $header['offset'], 0);
 $str = fread($f, $style_filesize - $header['offset']);
 fclose($f);
@@ -117,409 +113,67 @@ catch (Exception $error)
 	if(defined('XS_CLONING')) { @unlink($tmp_filename); }
 	xs_error($lang['xs_import_invalid_file'] . '<br /><br />' . $lang['xs_import_back']);
 }
-$files = array();	// complete list of files
-$list_data = array();	// result for list
-$dirs = array();	// complete list of directories
-$items = array();	// data
-foreach($archive_entries as $data)
+if (!$list_only)
 {
-	if($write_local)
+	require_once($phpbb_root_path . 'includes/functions_style_import.' . $phpEx);
+	require_once($phpbb_root_path . 'includes/functions_style_import_files.' . $phpEx);
+	$import_error = false;
+	try
 	{
-		$save_filename = $write_local_dir . $data['filename'];
+		$publisher = $write_local ? new PhpbbStyleImportLocal(rtrim($write_local_dir, '/\\')) : new PhpbbStyleImportFtp($ftp);
+		$publish = function ($guard) use ($publisher, $header, $archive_entries, $str) {
+			phpbb_style_import_publish($publisher, $header['template'], $archive_entries, $str, $guard);
+		};
+		$installed = phpbb_style_import($db, $HTTP_POST_VARS, $header, $archive_entries, $str, $publish);
 	}
-	else
-	{
-		$pos1 = strrpos($data['filename'], '/');
-		if($pos1)
-		{
-			$data['dir'] = substr($data['filename'], 0, $pos1);
-			$data['file'] = substr($data['filename'], $pos1 + 1);
-		}
-		else
-		{
-			$data['dir'] = '';
-			$data['file'] = $data['filename'];
-		}
-	}
-	if($data['typeflag'] === 5)
-	{
-		if($write_local && !$list_only)
-		{
-			if(!xs_create_dir($save_filename))
-			{
-				xs_error($lang['xs_import_invalid_file'] . '<br /><br />' . $lang['xs_import_back']);
-			}
-		}
-	}
-	$contents = $data['size'] > 0 ? substr($str, $data['offset'], $data['size']) : '';
-	$data['tmp'] = '';
-	// adding to list
-	$is_file = true;
-	if(intval($data['typeflag']) == 5)
-	{
-		$is_file = false;
-		if($data['filename'])
-		{
-			$dirs[] = $data['filename'];
-		}
-	}
-	else
-	{
-		if($data['filename'])
-		{
-			if(!$list_only)
-			{
-				if($write_local)
-				{
-					$res = xs_write_file($save_filename, $contents);
-					if(!$res)
-					{
-						if(defined('XS_CLONING'))
-						{
-							@unlink($tmp_filename);
-						}
-						xs_error(str_replace('{FILE}', $save_filename, $lang['xs_error_cannot_create_file']) . '<br /><br />' . $lang['xs_import_back']);
-					}
-				}
-				else
-				{
-					// write to temporary file
-					$data['tmp'] = @tempnam(XS_TEMP_DIR, 'xs_import_');
-					if($data['tmp'] === false || @file_put_contents($data['tmp'], $contents, LOCK_EX) !== strlen($contents))
-					{
-						if($data['tmp'] !== false)
-						{
-							@unlink($data['tmp']);
-						}
-						if(defined('XS_CLONING'))
-						{
-							@unlink($tmp_filename);
-						}
-						xs_error(str_replace('{FILE}', XS_TEMP_DIR, $lang['xs_error_cannot_create_tmp']) . '<br /><br />' . $lang['xs_import_back']);
-					}
-				}
-			}
-			elseif(!empty($get_file) && $get_file === $data['filename'])
-			{
-				// show contents of file
-				$f = $data['filename'];
-				$ext = strtolower(substr($f, strlen($f) - 4));
-				if(empty($HTTP_GET_VARS['get_content']) && xs_in_array($ext, $text_types))
-				{
-					// show as text
-					$str = '<div align="left">' . $lang['xs_import_list_contents'] . $f . ' [<a href="' . append_sid('xs_import.' . $phpEx . '?list=1&import=' . urlencode($filename) . '&get_file=' . urlencode($f) . '&get_content=1') . '">' . $lang['xs_import_download_lc'] . '</a>]<br /><br />';
-					$str .= '<textarea cols="120" rows="30" style="width: 100%">' . htmlspecialchars($contents, ENT_QUOTES, 'UTF-8') . '</textarea>';
-					$str .= '</div>';
-					xs_message($lang['Information'], $str);
-				}
-				else
-				{
-					$do_download = false;
-					$content_type = '';
-					if(empty($HTTP_GET_VARS['get_content']))
-					{
-						if($ext === '.gif')
-						{
-							$content_type = 'image/gif';
-						}
-						elseif($ext === '.jpg' || $ext === '.jpe' || $ext === 'jpeg')
-						{
-							$content_type = 'image/jpeg';
-						}
-						elseif($ext === '.png')
-						{
-							$content_type = 'image/png';
-						}
-						else
-						{
-							$do_download = true;
-						}
-					}
-					else
-					{
-						$do_download = true;
-					}
-					xs_download_file($do_download ? basename($f) : '', $contents, $content_type);
-					xs_exit();
-				}
-			}
-			else
-			{
-				$list_data[$data['filename']] = $data;
-			}
-			$files[] = $data['filename'];
-		}
-	}
-	if($is_file)
-	{
-		$items[] = $data;
-	}
-}
-if($list_only)
-{
-	// show list of files. used for debug.
-	$str = '<div align="left">';
-	// main data
-	$str .= $lang['xs_import_list_filename'] . htmlspecialchars($header['filename'], ENT_QUOTES, 'UTF-8') . '<br />';
-	$str .= $lang['xs_import_list_template'] . htmlspecialchars($header['template'], ENT_QUOTES, 'UTF-8') . '<br />';
-	$str .= $lang['xs_import_list_comment'] . htmlspecialchars($header['comment'], ENT_QUOTES, 'UTF-8') . '<br />';
-	$str .= $lang['xs_import_list_styles'] . htmlspecialchars(implode(', ', $header['styles']), ENT_QUOTES, 'UTF-8') . '<br />';
-	ksort($list_data);
-	$str .= '<br />' . str_replace('{NUM}', count($list_data), $lang['xs_import_list_files']) . '<br />';
-	$str .= '<table border="0" cellspacing="0" cellpadding="1" align="left">';
-	foreach($list_data as $var => $value)
-	{
-		$str .= '<tr><td>' . htmlspecialchars($value['filename'], ENT_QUOTES, 'UTF-8') . '</td><td>';
-		if($value['size'] > 0)
-		{
-			$ext = strtolower(substr($var, strlen($var) - 4));
-			if(xs_in_array($ext, $text_types) || xs_in_array($ext, $img_types))
-			{
-				$str .= '[<a href="' . append_sid('xs_import.' . $phpEx . '?list=1&import=' . urlencode($filename) . '&get_file=' . urlencode($var)) . '">' . $lang['xs_import_view_lc'] . '</a>] ';
-			}
-			$str .= '[<a href="' . append_sid('xs_import.' . $phpEx . '?list=1&import=' . urlencode($filename) . '&get_file=' . urlencode($var)) . '&get_content=1">' . $lang['xs_import_download_lc'] . '</a>] ';
-		}
-		$str .= str_replace('{NUM}', $value['size'], $lang['xs_import_file_size']) . '</td></tr>';
-	}
-	$str .= '</table>';
-	$str .= '</div>';
-	xs_message($lang['Information'], $str);
-}
-$str = '';
-if(!$write_local)
-{
-	//
-	// Generate actions list
-	//
-	$actions = array();
-	// chdir to template directory
-	$actions[] = array(
-			'command'	=> 'chdir',
-			'dir'		=> 'templates'
-		);
-	// create directory with template name
-	$actions[] = array(
-			'command'	=> 'mkdir',
-			'dir'		=> $header['template'],
-			'ignore'	=> true
-		);
-	// change directory
-	$actions[] = array(
-			'command'	=> 'chdir',
-			'dir'		=> $header['template']
-		);
-	// create all directories and upload all files
-	$actions[] = array(
-			'command'	=> 'exec',
-			'list'		=> generate_actions_dirs()
-		);
-	$ftp_log = array();
-	$ftp_error = '';
-	$res = ftp_myexec($actions);
-/*	echo "<!--\n\n";
-	echo "\$actions dump:\n\n";
-	print_r($actions);
-	echo "\n\n\$ftp_log dump:\n\n";
-	print_r($ftp_log);
-	echo "\n\n -->"; */
-	// remove temporary files
-	for($i=0; $i<count($items); $i++)
-	{
-		if(!empty($items[$i]['tmp']))
-		{
-			@unlink($items[$i]['tmp']);
-		}
-	}
-	if(!$res)
-	{
-		if(defined('XS_CLONING'))
-		{
-			@unlink($tmp_filename);
-		}
-		xs_error($ftp_error . '<br /><br />' . $lang['xs_import_back']);
-	}
+	catch (Exception $error) { $import_error = true; }
+	catch (Error $error) { $import_error = true; }
+	finally { if (defined('XS_CLONING')) { @unlink($tmp_filename); } }
+	if ($import_error) { xs_error($lang['xs_import_failed'] . '<br /><br />' . $lang['xs_import_back']); }
+	xs_message($lang['Information'], $lang[$installed ? 'xs_import_installed' : 'xs_import_uploaded'] . '<br /><br />' . $lang['xs_import_back']);
 }
 
-//
-// Check if we need to install style
-//
-$total = isset($HTTP_POST_VARS['total']) && is_scalar($HTTP_POST_VARS['total']) ? intval($HTTP_POST_VARS['total']) : 0;
-$total = max(0, min(XS_MAX_ITEMS_PER_STYLE, count($header['styles']), $total));
-$default = isset($HTTP_POST_VARS['import_default']) && is_scalar($HTTP_POST_VARS['import_default']) && strlen((string) $HTTP_POST_VARS['import_default']) ? intval($HTTP_POST_VARS['import_default']) : -1;
-$install = array();
-$default_name = '';
-for($i=0; $i<$total; $i++)
+// Preview never opens a writer/FTP connection or creates a directory.
+$list_data = array();
+foreach ($archive_entries as $data)
 {
-	$tmp = empty($HTTP_POST_VARS['import_install_'.$i]) ? 0 : 1;
-	if($tmp)
+	if ($data['typeflag'] !== 0) { continue; }
+	$f = $data['filename']; $ext = strtolower(substr($f, -4));
+	if (!empty($get_file) && $get_file === $f)
 	{
-		$set_default = $default == $i ? 1 : 0;
-		$tmp_name = $header['styles'][$i];
-		if($tmp_name)
+		$contents = substr($str, $data['offset'], $data['size']);
+		if (empty($HTTP_GET_VARS['get_content']) && xs_in_array($ext, $text_types))
 		{
-			$install[] = $tmp_name;
-			if($set_default)
-			{
-				$default_name = $tmp_name;
-			}
+			$html = '<div align="left">' . $lang['xs_import_list_contents'] . htmlspecialchars($f, ENT_QUOTES, 'UTF-8') .
+				' [<a href="' . append_sid('xs_import.' . $phpEx . '?list=1&import=' . urlencode($filename) . '&get_file=' . urlencode($f) . '&get_content=1') . '">' . $lang['xs_import_download_lc'] . '</a>]<br /><br />';
+			$html .= '<textarea cols="120" rows="30" style="width: 100%">' . htmlspecialchars($contents, ENT_QUOTES, 'UTF-8') . '</textarea></div>';
+			xs_message($lang['Information'], $html);
 		}
+		$types = array('.gif'=>'image/gif','.jpg'=>'image/jpeg','.jpe'=>'image/jpeg','jpeg'=>'image/jpeg','.png'=>'image/png');
+		$inline = empty($HTTP_GET_VARS['get_content']) && isset($types[$ext]);
+		xs_download_file($inline ? '' : basename($f), $contents, $inline ? $types[$ext] : '');
+		xs_exit();
 	}
+	$list_data[$f] = $data;
 }
-if(!count($install))
+$html = '<div align="left">';
+$html .= $lang['xs_import_list_filename'] . htmlspecialchars($header['filename'], ENT_QUOTES, 'UTF-8') . '<br />';
+$html .= $lang['xs_import_list_template'] . htmlspecialchars($header['template'], ENT_QUOTES, 'UTF-8') . '<br />';
+$html .= $lang['xs_import_list_comment'] . htmlspecialchars($header['comment'], ENT_QUOTES, 'UTF-8') . '<br />';
+$html .= $lang['xs_import_list_styles'] . htmlspecialchars(implode(', ', $header['styles']), ENT_QUOTES, 'UTF-8') . '<br />';
+ksort($list_data);
+$html .= '<br />' . str_replace('{NUM}', count($list_data), $lang['xs_import_list_files']) . '<br />';
+$html .= '<table border="0" cellspacing="0" cellpadding="1" align="left">';
+foreach ($list_data as $data)
 {
-	if(defined('XS_CLONING'))
+	$f = $data['filename']; $ext = strtolower(substr($f, -4));
+	$html .= '<tr><td>' . htmlspecialchars($f, ENT_QUOTES, 'UTF-8') . '</td><td>';
+	if ($data['size'] > 0)
 	{
-		@unlink($tmp_filename);
+		$url = 'xs_import.' . $phpEx . '?list=1&import=' . urlencode($filename) . '&get_file=' . urlencode($f);
+		if (xs_in_array($ext, $text_types) || xs_in_array($ext, $img_types)) { $html .= '[<a href="' . append_sid($url) . '">' . $lang['xs_import_view_lc'] . '</a>] '; }
+		$html .= '[<a href="' . append_sid($url . '&get_content=1') . '">' . $lang['xs_import_download_lc'] . '</a>] ';
 	}
-	xs_message($lang['Information'], $lang['xs_import_uploaded'] . '<br /><br />' . $lang['xs_import_back']);
+	$html .= str_replace('{NUM}', $data['size'], $lang['xs_import_file_size']) . '</td></tr>';
 }
-//
-// Get list of installed styles
-//
-$tpl = $header['template'];
-$sql = "SELECT themes_id, style_name FROM " . THEMES_TABLE . " WHERE template_name='" . xs_sql($tpl) . "'";
-if(!$result = $db->sql_query($sql))
-{
-	if(defined('XS_CLONING'))
-	{
-		@unlink($tmp_filename);
-	}
-	xs_error($lang['xs_import_notinstall'] . '<br /><br />' . $lang['xs_import_back']);
-}
-$style_rowset = $db->sql_fetchrowset($result);
-// run theme_info.cfg
-$data = xs_get_themeinfo($tpl);
-if(!@count($data))
-{
-	if(defined('XS_CLONING'))
-	{
-		@unlink($tmp_filename);
-	}
-	xs_error($lang['xs_import_notinstall2'] . '<br /><br />' . $lang['xs_import_back']);
-}
-// install styles
-$default_id = 0;
-for($i=0; $i<count($install); $i++)
-{
-	$style_name = $install[$i];
-	$style_data = false;
-	// find entry in theme_info.cfg
-	for($j=0; $j<count($data); $j++)
-	{
-		if($data[$j]['style_name'] === $style_name)
-		{
-			$style_data = $data[$j];
-		}
-	}
-	// check if already installed
-	$installed = 0;
-	for($j=0; $j<count($style_rowset); $j++)
-	{
-		if($style_rowset[$j]['style_name'] === $style_name)
-		{
-			$installed = $style_rowset[$j]['themes_id'];
-		}
-	}
-	// install/update
-	if(empty($style_data['style_name']) || empty($style_data['template_name']))
-	{
-		if(defined('XS_CLONING'))
-		{
-			@unlink($tmp_filename);
-		}
-		xs_error(str_replace('{STYLE}', htmlspecialchars($style_name, ENT_QUOTES, 'UTF-8'), $lang['xs_import_notinstall3']) . '<br /><br />' . $lang['xs_import_back']);
-	}
-	if($installed)
-	{
-		// update
-		$sql = '';
-		foreach($style_data as $var => $value)
-		{
-			if($sql)
-			{
-				$sql .= ', ';
-			}
-			$sql .= xs_sql($var) . " = '" . xs_sql($value) . "'";
-		}
-		$sql = "UPDATE " . THEMES_TABLE . " SET " . $sql . " WHERE themes_id = '{$installed}'";
-	}
-	else
-	{
-		// install
-		$sql = "SELECT MAX(themes_id) AS total FROM " . THEMES_TABLE;
-		if ( !($result = $db->sql_query($sql)) )
-		{
-			if(defined('XS_CLONING'))
-			{
-				@unlink($tmp_filename);
-			}
-			xs_error($lang['xs_import_notinstall4'] . '<br /><br />' . $lang['xs_import_back']);
-		}
-		if ( !($row = $db->sql_fetchrow($result)) )
-		{
-			if(defined('XS_CLONING'))
-			{
-				@unlink($tmp_filename);
-			}
-			xs_error($lang['xs_import_notinstall4'] . '<br /><br />' . $lang['xs_import_back']);
-		}
-		$installed = $row['total'] + 1;
-		$style_data['themes_id'] = $installed;
-		$sql1 = $sql2 = '';
-		foreach($style_data as $var => $value)
-		{
-			if($sql1)
-			{
-				$sql1 .= ', ';
-				$sql2 .= ', ';
-			}
-			$sql1 .= xs_sql($var);
-			$sql2 .= "'" . xs_sql($value) . "'";
-		}
-		$sql = "INSERT INTO " . THEMES_TABLE . " (" . $sql1 . ") VALUES (" . $sql2 . ")";
-	}
-	if ( !($result = $db->sql_query($sql)) )
-	{
-		if(defined('XS_CLONING'))
-		{
-			@unlink($tmp_filename);
-		}
-		xs_error($lang['xs_import_notinstall5'] . '<br /><br />' . $lang['xs_import_back']);
-	}
-	if($default_name === $style_name)
-	{
-		$sql = "UPDATE " . CONFIG_TABLE . " SET config_value='{$installed}' WHERE config_name='default_style'";
-		$board_config['default_style'] = $installed;
-		if(!$db->sql_query($sql))
-		{
-			if(defined('XS_CLONING'))
-			{
-				@unlink($tmp_filename);
-			}
-			xs_error($lang['xs_import_notinstall5'] . '<br /><br />' . $lang['xs_import_back']);
-		}
-	}
-}
-if(defined('XS_CLONING'))
-{
-	@unlink($tmp_filename);
-}
-if(count($install) && defined('XS_MODS_CATEGORY_HIERARCHY210'))
-{
-	// recache themes table
-	if ( empty($themes) )
-	{
-		$themes = new themes();
-	}
-	if ( !empty($themes) )
-	{
-		$themes->read(true);
-	}
-}
-if(count($install) && defined('XS_MODS_CATEGORY_HIERARCHY'))
-{
-	cache_themes();
-}
-xs_message($lang['Information'], $lang['xs_import_installed'] . '<br /><br />' . $lang['xs_import_back']);
-
-?>
+xs_message($lang['Information'], $html . '</table></div>');